Remove state container field from generated Component class
Summary:
Part of making Component immutable.
To remove the StateContainer field out of the Component, we start by removing it from the generated code. In a later diff, we'll have a flag to move it to the LayoutStateContext and access it by key.

Since we're currently comparing StateContainers in isEquivalentTo by reflection on the declared fields of the generated component, we'll need to explicitly compare the StateContainers since they're no longer fields on the subclassed Component.

Reviewed By: adityasharat

Differential Revision: D25095196

fbshipit-source-id: afe45b7fb97b91e165eff778d1605eba295940a8