Allow null LayoutState for sync layouts with cancelable futures
Summary:
Right now this throws an exception if we get a null LayoutState from a sync layout calculation - the assumption here is that only async LayoutStateFutures are cancellable.
However ComponentTree has an API to force a layout state future to be interrupted and ComponentWarmer uses that to cancel layouts that it knows will not be used at all, even if they're sync: https://fburl.com/diffusion/t4dj57we
So that's a valid scenario which we should allow and not crash on.

Reviewed By: Katalune

Differential Revision: D25467691

fbshipit-source-id: b33d84dc65db7bbd14a0692f3af9707ed22e86ce