Reset scroll position of horizontal scroll view on unmount (#731)
Summary:
If a horizontal scroll view is recycled, the "new" view will share the scroll position of the "old" view.

## Changelog

This change resets the scroll position of horizontal scroll views on `unmount` to prevent this recycling bug.

Pull Request resolved: https://github.com/facebook/litho/pull/731

Test Plan: Verified locally that this one-line change fixes the observed bug. Happy to add further testing if it would be useful.

Reviewed By: astreet

Differential Revision: D25585826

Pulled By: adityasharat

fbshipit-source-id: 85246bfe329536cbc505938990714840c30d30fb