Don't accept style in KComponent ctor
Summary: My reasoning here is that 1) I don't want to tie us to the class inheritance hierarchy 2) The dev still needs to manually declare they accept style anyway, so it's not much to specify where to put it. This means devs will consistently specify where they want a style applied, the same as React.js.

Reviewed By: colriot

Differential Revision: D24794932

fbshipit-source-id: a6069522a28c4db20a26d95e5e799b290dd4502c