Add initial litho onboarding
Summary:
The code samples live in the sample app under a top level Tutorial (we can move to Internal Debugging Samples if this is cluttering things).

The general idea is each step introduces a new concept and then calls out the new things you should take away in "Key Points".

There should maybe be separate pages and should definitely have images eventually, but we can add those once we're happy with the flow and code.

Reviewed By: colriot

Differential Revision: D25620251

fbshipit-source-id: e25399c841f61736247ba4bd2d478ff8da683eb3