Ignore duplicate transition keys in the same layout
The existing implementation is throwing an error if there are duplicate,
non autogenerated transition keys in the same layout. This change
exposes a new configuration parameter in ComponentsConfiguration to
ignore duplicate transition keys in layout and sets it to true by
default. It uses this parameter within LayoutState to not throw errors
if a duplicate transition key is detected, but ignore it instead.