TextInputSpec: Move the cursor to the end when the text is set (#732)
Summary:
This change ensures that whenever the text is set on a TextInput
component, the cursor is always placed at the end of the given text.
This allows users to continue inserting text at the end if so required

## Changelog

When text is set in the TextInput component, the cursor moves at the end.

Pull Request resolved: https://github.com/facebook/litho/pull/732

Test Plan: Added extra assertions in TextInputSpecTest

Reviewed By: colriot

Differential Revision: D25637975

Pulled By: adityasharat

fbshipit-source-id: a9a547255f4794f5fc0612528d18c66d26501e1a