Fix crashes caused by measuring of the Litho TextInputSpec (#737)
Summary:
Litho TextInputSpec replaces the deprecated EditTextSpec, however there
was a fix specific for setBackground in EditTextSpec that was not in
TextInputSpec. This change is copying over the fix from the deprecated
EditableTextSpec into the new TextInputSpec.

## Changelog

Reintroduce a fix into the new TextInputSpec from the deprecated EditableTextSpec
to resolve potential race conditions.

Pull Request resolved: https://github.com/facebook/litho/pull/737

Test Plan:
Verified locally that re-introducing this fix into TextInputSpec fixes the crashes
that we observed.

Reviewed By: colriot

Differential Revision: D25637996

Pulled By: adityasharat

fbshipit-source-id: f2832b2d715ba5e99ba9c633023b5c7654eba28b