Provide the ability to manually set initRangeSize in RecyclerBinder
At the moment the RecyclerBinder computes the layout synchronously on
the main thread to derive the rangeSize. This change exposes a setter
in the RecyclerBinder.Builder that allows users to set an expected
initial rangeSize and use that to derive the initRange rather than
measure the first component on the main thread.

This is a performance optimization, that is by default turned off and
only used when the initRangeSize parameter is explicitly set in the
RecyclerBinder builder.