Fix support for disappear transitions
The current setupTransitions() implementation never looks at any
transitions defined on the previous LayoutState when creating
AnimationBindings. This can be an issue for disappear transitions
because the new LayoutState will no longer contain any disappearing
items.

This change adds logic in the TransitionManager to first create
AnimationBindings for the previous LayoutState, then create
AnimationBindings for the new LayoutState and finally merge the two.