Add on scroll start / end detection support in ScrollableContainer
This change adds support for scroll start / end detection in
ScrollableContainer. It determines scroll status under the two
circumstances below:

1) For non-fling gestures, the ACTION_UP event indicates scroll end.
2) For fling gestures, the ScrollDetector continously checks if
onScroll() is called for each onDrawe event; If yes, then the
ScrollableContainer is currently scrolling. Otherwise scrolling is
stopped. This is based on the fact that onDraw() method always calls
onScrollChanged() internally.