Fix corner cases for scroll start/stop callbacks
This change fixes some corner cases for handling on scroll start / stop
callbacks in ScrollableContainer:

1) When the ScrollableContainer is within a RecyclerView, the TOUCH_DOWN
event is consumed and only TOUCH_MOVE and TOUCH_UP is available. In this
case the detector cannot notify the scroll start / stop callbacks
because it always waits for the TOUCH_DOWN event. This problem is fixed
by using both TOUCH_DOWN and TOUCH_MOVE to detect scrolling.

2) The TOUCH_CANCEL event should be treated as a TOUCH_UP.

3) During rematerialization, the ScrollStateDetector is recreated, which
is not necessary and is incorrect since the new instance has default
values for all local variables which might differ to the values held in
the previous instance (thus breaking the detection mechanism). In this
change we tie the lifecycle of ScrollStateDetector to the internal
scrollView which fixes this issue and is more performant overall.