Avoid mocking Android Views in DynamicPropsTest (#734)
Summary:
The testDynamicElevationApplied() test in DynamicPropsTest is currently
using Mockito to mock an Android View. This change removes mocking and
uses a fake LithoView extension instead to retrieve the set elevation.

## Changelog

Remove the need to mock an Android View in DynamicPropsTest

Pull Request resolved: https://github.com/facebook/litho/pull/734

Test Plan: This change only affects tests and is not user visible

Reviewed By: colriot

Differential Revision: D25637987

Pulled By: adityasharat

fbshipit-source-id: 2e5030e063a1af0a1b83359bf838b1dc02873915