Override getColorFilter() in BorderColorDrawable (#745)
Summary:
This change overrides getColorFilter() in BorderColorDrawable to provide
the ColorFilter set on the Paint if the latter is available or null
otherwise. This getColorFilter() implementation is symmetrical to the
existing setColorFilter() implementation.

## Changelog

Provides a symmetrical getColorFilter() implementation in BorderColorDrawable,
to match the existing implementation of setColorFilter().

Pull Request resolved: https://github.com/facebook/litho/pull/745

Test Plan:
We have manually verified that adding this getColorFilter() implementation
fixes an issue observed around the border color not being applied properly
during animations.

Reviewed By: mihaelao

Differential Revision: D25692339

Pulled By: adityasharat

fbshipit-source-id: 0f56725c46701e8c43e4e948634b87b87b26836e