Add support for OnItemTouchListener in RecyclerSpec (#739)
Summary:
An optional property has been added into RecyclerSpec so that users can
optionally provide an OnItemTouchListener. The given Listener is passed
into the underlying RecyclerView to be triggered when an item has been
touched.

## Changelog

Adds support for OnItemTouchListener in RecyclerSpec

Pull Request resolved: https://github.com/facebook/litho/pull/739

Test Plan:
Updated the RecyclerSpecTest to reflect the change and verified it is working
as expected locally.

Reviewed By: mihaelao

Differential Revision: D25692261

Pulled By: adityasharat

fbshipit-source-id: 39f019c848e0c51539050420ad73f5b500bbd1b4