Un-deprecate the estimatedViewPortCount API in RecyclerBinder
The estimatedViewPortCount API was added a few months back and was
initially marked as @Deprecated. This API has found legitimate uses in
certain scenarios, so it can be un-deprecated and added to the list of
public RecyclerBinder APIs, but with a comment offering a warming that
it should only be used under specific scenarios and not in the general
case.