Un-deprecate the estimatedViewPortCount API in RecyclerBinder (#749)
Summary:
The estimatedViewPortCount API was added a few months back and was
initially marked as Deprecated. This API has found legitimate uses in
certain scenarios, so it can be un-deprecated and added to the list of
public RecyclerBinder APIs, but with a comment offering a warming that
it should only be used under specific scenarios and not in the general
case.

## Changelog

Un-deprecate the estimatedViewPortCount API in RecyclerBinder

Pull Request resolved: https://github.com/facebook/litho/pull/749

Test Plan: N/A, as this commit is mainly affecting documentation.

Reviewed By: astreet

Differential Revision: D25802029

Pulled By: mihaelao

fbshipit-source-id: 3f1ff8d4b354d546c3e3e19d4718316a473ebaa8