Ignore duplicate transition keys in the same layout (#740)
Summary:
The existing implementation is throwing an error if there are duplicate,
non autogenerated transition keys in the same layout. This change
exposes a new configuration parameter in ComponentsConfiguration to
ignore duplicate transition keys in layout and sets it to true by
default. It uses this parameter within LayoutState to not throw errors
if a duplicate transition key is detected, but ignore it instead.

## Changelog

Ignores duplicate transition keys in layout by default, rather than throwing error.

Pull Request resolved: https://github.com/facebook/litho/pull/740

Test Plan: Verified locally that this change is working as intended.

Reviewed By: astreet

Differential Revision: D25823609

Pulled By: adityasharat

fbshipit-source-id: 0973356ac083cccc5069edf3b36a902ab6fb339e