Add Dimen interface
Summary: This diff changes a bit how we do dimensional values (dp, sp, px). We need some sort of interface to accept Px and Dp on dimen props, and they need to be able to convert to pixels. Note that I'm not distinguishing between props that take Dp/Px (dimensions) vs. Sp/Px (text) vs. any of the three (positions). We can add that later if we thinks it's valuable, imo it just complicates things.

Reviewed By: strulovich

Differential Revision: D25848116

fbshipit-source-id: e935cbb77a0fa391c4ea1f9433016e4c7dc90a2e