Add ScopedComponentInfo class
Summary:
Introducing the ScopedComponentInfo class, which will contain all the layout-scoped stateful information that the Component currently hosts as fields.
For stateless components, the goal is to move all fields of the Component to a ScopedComponentInfo which will be hosted outside of the Component (by the LayoutStateContext - see diffs above).

Here, we're starting by replacing the StateContainer field of the Component with a ScopedComponentInfo field which wraps the StateContainer. In the next diffs, we'll experiment with keeping the ScopedComponentInfo on the LayoutStateContext (eventually we'll move out all the fields).

This seems easier to gate and move around than doing it field by field, let me know what you think.

Reviewed By: adityasharat

Differential Revision: D25197155

fbshipit-source-id: 626fa7ac20a53b96a9dfd3a97ae395abf0c32ee3