Get StateContainers from LayoutStateContext
Summary:
Keep a map of global key to ScopedComponentInfo in the LayoutStateContext and access the StateContainer from there instead of keeping it as a field on the Component.

There are cases where we only have access to the scoped ComponentContext and we need the StateContainer, so we need to do a map lookup to get it when we could keep it directly on the ComponentContext - I'd like to hear you opinion on this.

I would like to move towards keeping any layout-scoped stuff on the LayoutStateContext - eventually the ComponentContext becomes only a wrapper for the global key and the Android context.

Reviewed By: adityasharat

Differential Revision: D25197175

fbshipit-source-id: 6f60e73720f4b96b902ba496d6666b345556d54d