Set focusable false when using AX delegate
Summary:
Right now we are using a forked version of ExploreByTouchHelper for fb4a and ig4a which comments out the host.setFocusable(true) which is by default on the View. For Whatsapp we are not using the forked version and this works ok in general but when trying to add expresso tests that simulates a tab key down this will request focus on the view (even without ax enabled).
This change matches what we do in the forked ExploreByTouchHelper.

Reviewed By: pasqualeanatriello

Differential Revision: D25831887

fbshipit-source-id: adedf60101eabd7bcc31d350b35377ebf13632e9