New immutable style impl
Summary:
Reworks the Style implementation in preparation for moving more properties to it.

The API is staying the same except
1) You no longer need to import each style property separately
2) Everything now uses the Style.blah(...) API and not Style() directly
3) Setting position no longer automatically sets positionType

We are changing the implementation to try to reduce large object allocations and having to copy over all the fields each time a property is set.

The implementation is based on Style and StyleItem.
- A StyleItem knows how to apply a single attribute like width
- Style is an ordered, immutable collection of StyleItems

While adding a new StyleItem to a Style, a new Style is created pointing to the old Style plus the new StyleItem. This is similar to adding a new head to a linked list, where the value is the StyleItem and the next pointer is the old style. Similarly, when combining two Styles a new Style is returned pointing to both.

I recognize this implementation doesn't allow constant-time style attribute lookups -- we'll have to see how much that matters. There are definitely alternate implementations here with different "tradeoffs", e.g. using an immutable map and copying it each time (if we're concerned this matters, we need to measure it which we haven't put time into doing).

Last thing to note about this implementation: separating Style into StyleItems allows us to decentralize the usable Style attributes which is important given the extensibility we want to offer. Of course, we aren't actually taking advantage of it in this diff, namely because Litho's core is not yet set up to allow defining custom attributes (i.e. everything still uses CommonProps).

Reviewed By: strulovich

Differential Revision: D25802862

fbshipit-source-id: dd969d1fbff7e01a12da3f26a2709515fac22fd5