Ensure that a ComponentContext param is available for accessing state
Summary:
In code gen, when calling `getStateContainerImpl(<ComponentContext param>)` on delegate methods, we need to infer the name of the ComponentContext param to use it.
There were cases in code gen where this was overlooked and `null` was passed instead of the context name. This adds validation that `getStateContainerImpl` is always called with a context name and adds an error message with the name of the method where this is called if the validation fails.

To fix existing issues after adding the validation, `ComponentLifecycle.recordRenderData` signature changed to accept a ComponentContext which can be used in the generated code.

Reviewed By: adityasharat

Differential Revision: D25955611

fbshipit-source-id: 4b3b0417ea38b6a030fc14eefd4455620e08b21d