BaselineToCapHeight for renderCore text
Summary:
This diff adds the ability to manually control spacing from Cap height and from baseline. This is a requirement for CDS and it's something that is's small enough that it's ok to directly add to our text implementation.
I want to make sure this is actually what we need for CDS. thurn let me know what you think.

Reviewed By: thurn

Differential Revision: D25944762

fbshipit-source-id: 79c8591f32fee0345295ce514a3ddea7013c99cd