Adds @LooperMode LEGACY in tests that fail in PAUSED mode
The new Robolectric PAUSED LooperMode will become the default
in new versions of Robolectric. We have ran all tests against
the new PAUSED LooperMode and unfortunately some tests are
failing. For these tests we need to explicitly specify the
LooperMode as LEGACY to be future proof, so that they do not
start failing when the default changes to PAUSED in the future.