Remove HooksHandler
Summary: Removes the separate HooksHandler code. I feel pretty strongly that all ComponentTree state that outlives a single layout creation (which includes both State and hooks) should live on a single object, which currently is StateHandler. The fact that we had to duplicate so much code indicates that HooksHandler should probably have just gone on StateHandler instead of being a parallel object. As we add more hooks, I'm definitely open to refactoring the hooks state logic in StateHandler to accommodate hooks in general, but currently it doesn't seem necessary.

Reviewed By: colriot

Differential Revision: D26174566

fbshipit-source-id: fc99f6cfc1671575440f4279a76da65dd8864339