Simplify useState impl
Summary:
- Changes `by useState()` to `= useState()` (which makes it the same as useTreeProp).
- Changes updating state from having to update in a `updateState { }` block to calling `.update` on the State object.

I considered overriding `set()` on the value, but then we'd still need to add an `update` that takes a lambda. For now, this means if someone wants to update two state vars at once, they could trigger two state updates. In the short term, we will want people to store state that needs to update together in the same object so that there's one state update. In the longer term, we should collect updates while dispatching events to do at once and/or expose a call that implements batching. I think this is fine for now though.

Reviewed By: colriot

Differential Revision: D26174555

fbshipit-source-id: a9b249471e6330bc175281ba99ac57502b140c01