Release nested LithoViews when parent is released
Summary:
When a LithoView with nested children LithoViews gets released, we should notify the children LithoViews to release too.
Fixes issue with items in HorizontalScrollSpec/VerticalScrollSpec not getting a detach callback even if the parent gets it.

D25184714 specifically addresses the detach callback not being forward by HorizontalScroll to the children - if we decide to run this as a test I can comandeer that too.

Reviewed By: adityasharat

Differential Revision: D26100125

fbshipit-source-id: 41d0c326c6d75669daa0a3799fe0644486b708be