Move interstage props out of generated Component class
Summary:
Interstage props are currently members of the generated Component/Surface class.

This change introduces an interstage prop wrapper class: InterStagePropsContainer.java,  that's created similar to the StateContainter and is kept on the Component class.

The code generation will create the static InterStagePropsContainer implementation that holds the interstage props, but the Component will create and keep the InterStagePropsContainer instance, same as with the StateContainer.

In the next diff, the InterStagePropsContainer instance will be moved to the LayoutStateContext and out of the Component - this will be needed for Component immutability

Reviewed By: astreet

Differential Revision: D25498536

fbshipit-source-id: 046cd7cae570585683ff8d3b50209e1b128559f2