Change signature of copyInterStageImpl
Summary:
The current generated implementation of `copyInterStageImpl` accesses `getInterStagePropsContainerImpl()` to get the source and destination InterStagePropsContainers (to copy from and into).
In the next diff, we'll change the signature of `getInterStagePropsContainerImpl` to require a ComponentContext param, so that we can store the InterStagePropsContainer on the LayoutStateContext instead of the Component.
We can either change `copyInterStageImpl` to pass the scoped contexts of the source and destination Components, or pass the InterStagePropsContainers directly. I went with the second option, since it gives more flexibility in the future to decide where to keep the InterStagePropsContainers.

Reviewed By: adityasharat

Differential Revision: D25867594

fbshipit-source-id: d4746fdd148e2bc4ba57791867527fc0b2150b5a