Allow setting custom permitted annotations on @OnEvent method params
Summary:
There are SpecModels which have custom annotations that can be used in OnEvent methods, but they are not part of the litho package. This enables passing the list of allowed annotations for event param methods so validation can pass.

See next diff for how this is used in SurfaceSpecModels

Reviewed By: adityasharat

Differential Revision: D26312573

fbshipit-source-id: b17a2f381658935b8c3a13ef2652ced9f6300d62