Change signature of copyInterStageImpl
Summary:
reland D25867594 (https://github.com/facebook/litho/commit/e8773eaa38e47d3b8a6c0aac87f2ca1fb17d2337)

The current generated implementation of copyInterStageImpl accesses getInterStagePropsContainerImpl() to get the source and destination InterStagePropsContainers (to copy from and into).
In the next diff, we'll change the signature of getInterStagePropsContainerImpl to require a ComponentContext param, so that we can store the InterStagePropsContainer on the LayoutStateContext instead of the Component.
We can either change copyInterStageImpl to pass the scoped contexts of the source and destination Components, or pass the InterStagePropsContainers directly. I went with the second option, since it gives more flexibility in the future to decide where to keep the InterStagePropsContainers.

Reviewed By: adityasharat

Differential Revision: D26403106

fbshipit-source-id: 26ef46db9f790e48da82bbdf003255fefef14654