Moves some InternalNode mutation from collect results to InternalNode#freeze
Summary:
This diff moves some of the post layout calculation mutations of the result to `InternalNode#freeze`.

In principle mutations to the `LayoutResult` should not be allowed after layout calculation. `LayoutState#collectResults` mutates the A11Y and enabled property of results post the layout calculation.

When `InternalNode` stops being the `LayoutResult` the current call sites of these mutation will not work because these APIs are only available on the `InternalNode`.

Removes unnecessary hasNewLayout, and `markLayoutSeen` APIs.

Reviewed By: SidharthGuglani

Differential Revision: D26373731

fbshipit-source-id: fd2466133ea4ca1e5f33a0139aaaafa952832de5