Uses LithoLayoutResult in collect results
Summary:
This change set uses the LithoLayoutResult in collect results; this in effect means that collect results will operate on the outputs. A couple of place required a hard cast back to InternalNode; these are mutations, and will be moved out of collect results in upcoming diffs.

See D26373731 where ToDos are addressed.

Reviewed By: mihaelao

Differential Revision: D26368633

fbshipit-source-id: 40d0e5243a59034db8c4094ee4d139c416777908