Use source of current layout to decide future interruption
Summary:
if the LayoutStateFuture created for the current layout needs to wait for the result of a running layout, it's reassigned. So the future we check the source on doesn't match the source of the current layout calculation.  https://fburl.com/diffusion/fqytohy7.
https://fburl.com/diffusion/k1dma1yh
We can end up interrupting a layout that is not from a sync layout, but validate that the future was okay to interrupt using the actual source of the layout, which is different.

This change ensures that we always use the source of the current layout to perform checks. There's no need to keep the source on the LayoutStateFuture, it's a source of bugs.

Reviewed By: adityasharat

Differential Revision: D26485678

fbshipit-source-id: f0b48fa9d772f42faba5e0f7202eb762fbe53e63