Fix bindings being removed while looping through them.
Summary:
This should fix T85753905. I'm not 100% sure as I was not able to repro this.
The problem here is that when we loop through the bindings list we can end up removing more than one binding in each loop, and this can end up throwing index out of bounds.
The sequence is:
- Loop through the graph bindings.
- Call binding.notifyNodesHaveFinished() if all nodes have finished.
- Call an added listener onAllNodesFinished method.
- This can end calling stop() on a binding with several bindings (ParallelBinding for example).
- The stop() method in TransitionAnimationBinding will call deactivate() on the GraphBinding which will end up calling mDataFlowGraph.unregister(this);
- This unregister removes the binding from the mBindings list we are looping.

Tried reproducing this but couldn't.

Also changed the mBindings to a set as it makes more sense since we are just looping through the list and removing objects.

Reviewed By: astreet

Differential Revision: D26777764

fbshipit-source-id: 60e88a593c6a771df7559271ae3018d3f76425f8