Add role description to AccessibleClickableSpan
Summary:
This adds an optional roleDescription property to AccessibleClickableSpan.  When used, this replaces the default string used for the AccessibilityRole with a custom one, while retaining the functionality of any role used.

For example, setting the "Button" role with a roleDescription of "Link" will still allow navigation to the spannable via the "Controls" navigation mode (enabled by the use of the Button role), but will announce "Link" upon focus rather than "Button".

Reviewed By: ikenwoo

Differential Revision: D26742991

fbshipit-source-id: 6a83a9b6f607f90fd952495d91f6c04f919d8dfa