Adding interpolator and two callbacks for onFinish and onUpdate to the timing call
Summary:
$title

I extracted the listeners to the separate call to clean the code a little bit.

For demo app I added logs, but maybe I should add small num value (for update calls) inside the view or something else that would show the working calls without the need of logcat?

+ i remember seeing somewhere something about ?.let use at Facebook but coudn't find it.  Is it ok to use it?

Reviewed By: astreet

Differential Revision: D26691479

fbshipit-source-id: 8b2a5c25f7af685cc175fa825b8a50514317fd89