Fix adding bindings while looping them
Summary:
Same problem with this D26777764 (https://github.com/facebook/litho/commit/6361d7ed5ced4cd710c2514bca65bb76bfb2e6d6) diff, we are also registering bindings when finishing others. I was able to repro this when trying to comment in a story. There is a parallel binding inside a sequence and when the parallel finishes the sequence will register the next binding in the sequence.

Error:
{F451009156}

Fix:
{F451012102}

Reviewed By: astreet

Differential Revision: D26815185

fbshipit-source-id: 12fc2a1e839a2a4d440cf39a4959cac16e4d122e