Type DataDiffSection.create correctly, for all DataDiff
Summary:
We want D26883395, which lets RenderEvent be well-typed.

Unfortunately, it will fail immediately if you try to pass a non-typed RenderEvent into a method requiring a typed RenderEvent, and vice versa. This is good: it means, once we land D26883395, all pagination sections must be perfectly and exactly typed, and will have compilation failures otherwise.

But to get there, we need to type everything.

Reviewed By: adityasharat

Differential Revision: D27125653

fbshipit-source-id: 9cc231f71cc14a1a519a26d43ee5aa42edbba3ee