using layout state context from ComponentContext instead of passing it separately as parameter
Summary:
We pass LayoutStateContext to Layout.createAndMeasureComponent method and ComponentContext is also passed here. This is creating problem as now there are now two places from which LayoutStateContext can be accessed.

In this diff we are now not passing LayoutStateContext to Layout.createAndMeasureComponent and instead getting LayoutStateContext from ComponentContext itself making it simpler and avoiding issues where LayoutStateContext was not set on ComponentContext which were causing NPE crashes.

Reviewed By: adityasharat

Differential Revision: D27154004

fbshipit-source-id: bb903bb7c33139dcee7313cb557776f3ff4f11dd