Add OnCreateInitialState for SectionSpec
Summary:
Add one method completion, `OnCreateInitialState` to `SectionSpec`. This is almost the same as `OnCreateInitialState` for other specs. The difference is `SectionSpec`'s uses `SectionContext` instead of `ComponentContext`.

I have changed the way to look up for method completion. I added the prefix with dot separation to specify which method we want to use in case there are two different method signatures with the same name.

Reviewed By: Katalune

Differential Revision: D27193496

fbshipit-source-id: 9b2be3476706a76be52a1c7e5b17b8cf2769bee5