Capture partially resolved layout root in Output
Summary: #Eventually `Layout.mLayoutRoot`, and `Layout.create()` return type needs to be `LithoLayoutResult`; but an interrupted layout will be an `InternalNode`. In order to return an `InternalNode` from `Layout.create()` when the layout is interrupted this change set passes and Output which will contains the partially resolved `InternalNode` tree. In D26425613 the partially resolved layout is held in another i-var on `LayoutState`.

Reviewed By: mihaelao

Differential Revision: D26425611

fbshipit-source-id: 70917341e84b77cc8dd0879a436babf2ec6dc74a