Fixes bug in component bounds shown in Flipper
Summary:
Fixes bug in component bounds shown in Flipper.

Flipper needs to traverses up the layout result hierarchy to get the bounds of the component. Since nested tree do not have any parent, their top, and left always default to 0,0. This diff adds the `setParent()` api to `LithoLayoutResult` for `DebugComponent` to use.

This will be refactored out once we redo the integration with Flipper.

Reviewed By: mihaelao

Differential Revision: D27400459

fbshipit-source-id: 9790226a76dafbd3da9ca13e86c0ae0b8178b4ef