Don't unmount all when nested components update
Summary:
For nested tree Components, setComponentTree(null) is essentially not doing anything.
However now that we're also unmounting all when setting a null component tree, it means that when the parent of a nested component goes through a state update, we'll unmount the content of the nested component which is not what we want.

Reviewed By: adityasharat

Differential Revision: D27618704

fbshipit-source-id: a466a2033bdf22780213f8f6ddb920aaa96f9f8b