Allow a kotlin Style to be applied to a spec Component Builder
Summary: There are cases where a spec component is converted to Kotlin, but returns a component that hasn't been converted to Kotlin yet. In that case, the parent kotlin component will receive common props via a Style property (if it  defines one), but needs a way to apply those properties to the child spec component builder. This API allows the properties on that Style to be applied to the builder.

Reviewed By: colriot

Differential Revision: D27618153

fbshipit-source-id: 5ec566a2f61e59e204ca2b9e23f920da6f2611cc