Add Auto-Completed Method Signature
Summary: Litho context now becomes optional. The plugin needs to be updated to reflect it. This diff adds method signature to help developer able to choose which method they want to use. For example, in the next diff, there are 2 overloaded methods of `OnCreateLayout`; developer might not notice the change, so the plugin includes both of them. ~~This diff starts by adding method signatures to show which overloaded method they should use. ~~

Reviewed By: Katalune

Differential Revision: D27618379

fbshipit-source-id: fffe0958029cb653fcff41933a6563fdf42453fc