Fix ellipsizing with styled custom ellipsis text (#769)
Summary:
At the moment it seems that Litho is not measuring styled ellipsis text properly.
This PR creates a new text Layout for the custom ellipsis text and uses that to
calculate the text width and drive text truncation. It also caters for cases when
the text might be left aligned or have an RTL direction.

## Changelog

Fixes the way styled ellipsis text is measured.

Pull Request resolved: https://github.com/facebook/litho/pull/769

Test Plan: Unit tests have been added.

Reviewed By: astreet

Differential Revision: D27595380

Pulled By: colriot

fbshipit-source-id: aca79e12953f432c33944d440ff34df27b6cf74d