RecyclerBinder: set up a LithoLifecycleOwner for nested ComponentTrees
Summary:
Each ComponentTree will be subscribed to a RecyclerBinderLithoLifecycle.
The RecyclerBinderLithoLifecycle listens to changes in the RecyclerBinder's parent lifecycle. The RecyclerBinderLithoLifecycle changes when the parent's lifecycle changes, as well as go through its own state changes - when the ComponentTree goes out of the prepared range, the RecyclerBinderLithoLifecycle goes to the destroyed state.

Reviewed By: astreet

Differential Revision: D27428783

fbshipit-source-id: 88b88a2314755a488f02bcc1465d4cf131b1f880