LithoView subscribe to lifecycle API
Summary:
We've discussed about exposing an API on the LithoView to subscribe to lifecycle state changes.

The goal is to hide away the APIs which were previously used to react to state changes and move to a model where the user only informs the LithoView that a state change has occurred, and the LithoView will decide how to handle it.

This api enables to subscribe the LithoView as observer to:
- a Lifecycle object: this will subscribe the LithoView to observe all state changes
- a LifecycleRegistry object: this will allow the user to decide which state changes the subscribed LithoView is notified about.

Visibility changes that aren't tied to an existing lifecycle state change require special handling, so the user can manually inform the subscribed LithoViews when a visibility change occurs (for cases like ViewPager tab change or adding a fragment on top of another)

Reviewed By: astreet

Differential Revision: D26982485

fbshipit-source-id: ede5de678379baddff577296d5c9415562cd891a