Add content description span to Litho
Summary:
Right now, there is no generic span type that allows you to replace some span of content with a contentDescription for accessibility.

You can do this with AccessibleClickableSpan, but only if you want the element to be clickable.  In API 30+ you can also do this with ReplacementSpan, but this won't work on lower API levels.

The primary use case here is to be able to give contentDescriptions to ImageSpans, but by separating this out into its own spannable, it means it can be used for various other cases, such as giving a better description when the text itself is referencing something else on screen.  For example, the following text:

"Click the button above to continue"

Would not be helpful for a blind user who may not be able to easily find the button, or know *which* button is being described.  You could instead replace "button above" with a ContentDescriptionSpan of "next step button" so that a screen reader will announce:

"Click the next step button to continue".

{F547018175}

Reviewed By: nicous

Differential Revision: D27378460

fbshipit-source-id: 4af0a16e0ec45fcf3972e375abe21b06f24b91a0