Changes the signature of InternalNode#calculateLayout()
Summary:
`InternalNode` will not contain a YogaNode, therefore the width and height cannot be set on the root node using the `InternalNode.setStyle*FromSpec` APIs.

This change set moves the logic to set the width, height, and, layout direction from `Layout.java` to the implementation of `InternalNode.calculateLayout()`.

In order to do this the signature needed to change:

* Accept the width and height spec instead of the raw width and height.
* Pass the `ComponentContext` to set the correct layout direction.

These changes will also become relevant once the input only InternalNode is introduced in the diffs above.

Reviewed By: mihaelao

Differential Revision: D27619922

fbshipit-source-id: c1a09fa29efc32178515a2a6e0ffdefea37e7442