Refactors InternalNode apply attributes implementation
Summary:
**API only change**

Since the input only internal node will lazily apply the LayoutProps to YogaNode from the style attributes the InternalNode apply attributes API needs to accept the `int defStyleAttr,  int defStyleRes` instead of a `TypedArray`.

This change set makes that API change for `DefaultInternalNode` to reduce the size of the future change set.

Reviewed By: mihaelao

Differential Revision: D27647363

fbshipit-source-id: 772d131c486b69c7f87b7b3f22cc1797425a7577