Fixing removal of dynamic props by SpecModelImpl
Summary:
Due to the fact that we were removing props from the list by the size of dynamic props we always were removing things from the end of the list.
    final int basePropsSize = rawProps.size() - diffPropModels.size();

In the scenario that we have shouldUpdate and onBindDynamic Value, the dynamic value is the last one that is bing removed.

Reviewed By: astreet

Differential Revision: D27767452

fbshipit-source-id: 3aa7af1d2497c1450d42e4b0175c2f05d40afb35