Fix bounds used to release moved items
Summary:
`mEstimatedViewportCount` is only an estimation, so if we rely on this to decide to release trees we may incorrectly release trees which are in the viewport. RecyclerBinder will still force items to be computed and rendered if they're outside of the precomputed range but need to be bound to a view.

Use the last visible position instead to determine whether an item is not in the precomputed range and is also not part of the visible range.

Reviewed By: adityasharat

Differential Revision: D27856160

fbshipit-source-id: 852703466c8598a84e0081dc8bfbdb55f8fa524b