Fix Spec Template for Kotlin (2nd Try)
Summary:
Found simpler solution for D27709474

 ---
It is [reported that Litho template is misbehaving](https://fb.workplace.com/groups/kotlinandroidqa/permalink/2918077255138165/). The investigation found that the implementation of Java and Kotlin is different in IntelliJ 2019.3; For Java file, file name is changed to match class name inside the file, while in Kotlin nothing changes.

This diff directly forces the name to add the suffix at creation time and changes the Kotlin template files to exclude the suffix in the files

Also, specific test cases are added

Reviewed By: Katalune

Differential Revision: D27856294

fbshipit-source-id: 45afed2d72f5f544f74abcfcb5be496d89e4f0ed